#1 Refactor package structure and add #4 input parsing test
Reorganized `org.example` to `aau.webcrawler` for clearer project context and consistent naming. Added `WebCrawlerInputTest` to ensure correct input argument parsing.
