+++ src/main/java/aau/webcrawler/WebsiteAnalyzer.java
@@ -1,4 +1,4 @@
-package org.example;
+package aau.webcrawler;
 
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
