+++ src/test/java/aau/webcrawler/MarkdownFileWriterTest.java
@@ -1,4 +1,5 @@
-import org.example.MarkdownFileWriter;
+package aau.webcrawler;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
